+++ custom_components/plugwise/models.py
@@ -23,21 +23,28 @@ from homeassistant.const import (
     UnitOfVolume,
     UnitOfVolumeFlowRate,
 )
-from plugwise import DeviceData
+from plugwise import SmileBinarySensors, SmileSensors, SmileSwitches
 
 
 @dataclass
 class PlugwiseSensorBaseMixin:
-    """Mixin for required Plugwise sensor base description keys."""
+    """Mixin for required Plugwise sensor description keys."""
 
-    value_fn: Callable[[DeviceData], float | int]
+    value_fn: Callable[[SmileSensors], float | int]
+
+
+@dataclass
+class PlugwiseBinarySensorBaseMixin:
+    """Mixin for required Plugwise binary_sensor description keys."""
+
+    value_fn: Callable[[SmileBinarySensors], bool]
 
 
 @dataclass
 class PlugwiseSwitchBaseMixin:
-    """Mixin for required Plugwise switch base description keys."""
+    """Mixin for required Plugwise switch description keys."""
 
-    value_fn: Callable[[DeviceData], bool]
+    value_fn: Callable[[SmileSwitches], bool]
 
 
 @dataclass
@@ -54,7 +61,7 @@ class PlugwiseSwitchEntityDescription(SwitchEntityDescription, PlugwiseSwitchBas
 
 @dataclass
 class PlugwiseBinarySensorEntityDescription(
-    BinarySensorEntityDescription, PlugwiseSwitchBaseMixin
+    BinarySensorEntityDescription, PlugwiseBinarySensorBaseMixin
 ):
     """Describes Plugwise binary sensor entity."""
 
@@ -68,7 +75,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["setpoint"],
+        value_fn=lambda data: data["setpoint"],
     ),
     PlugwiseSensorEntityDescription(
         key="cooling_setpoint",
@@ -76,7 +83,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["setpoint_high"],
+        value_fn=lambda data: data["setpoint_high"],
     ),
     PlugwiseSensorEntityDescription(
         key="heating_setpoint",
@@ -84,7 +91,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["setpoint_low"],
+        value_fn=lambda data: data["setpoint_low"],
     ),
     PlugwiseSensorEntityDescription(
         key="temperature",
@@ -92,7 +99,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["temperature"],
+        value_fn=lambda data: data["temperature"],
     ),
     PlugwiseSensorEntityDescription(
         key="intended_boiler_temperature",
@@ -100,14 +107,14 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["intended_boiler_temperature"],
+        value_fn=lambda data: data["intended_boiler_temperature"],
     ),
     PlugwiseSensorEntityDescription(
         key="temperature_difference",
         translation_key="temperature_difference",
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.KELVIN,
-        value_fn=lambda data: data["sensors"]["temperature_difference"],
+        value_fn=lambda data: data["temperature_difference"],
         icon="mdi:temperature-kelvin",
     ),
     PlugwiseSensorEntityDescription(
@@ -115,7 +122,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="outdoor_temperature",
         device_class=SensorDeviceClass.TEMPERATURE,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["outdoor_temperature"],
+        value_fn=lambda data: data["outdoor_temperature"],
         suggested_display_precision=1,
     ),
     PlugwiseSensorEntityDescription(
@@ -124,7 +131,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["outdoor_air_temperature"],
+        value_fn=lambda data: data["outdoor_air_temperature"],
         suggested_display_precision=1,
     ),
     PlugwiseSensorEntityDescription(
@@ -133,7 +140,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["water_temperature"],
+        value_fn=lambda data: data["water_temperature"],
     ),
     PlugwiseSensorEntityDescription(
         key="return_temperature",
@@ -141,21 +148,21 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["return_temperature"],
+        value_fn=lambda data: data["return_temperature"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed",
         translation_key="electricity_consumed",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_consumed"],
+        value_fn=lambda data: data["electricity_consumed"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_produced",
         translation_key="electricity_produced",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_produced"],
+        value_fn=lambda data: data["electricity_produced"],
         entity_registry_enabled_default=False,
     ),
     # Does not exist in Core - related to P1v2
@@ -164,7 +171,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="electricity_consumed_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_point"],
+        value_fn=lambda data: data["electricity_consumed_point"],
     ),
     # Does not exist in Core
     PlugwiseSensorEntityDescription(
@@ -172,35 +179,35 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="electricity_produced_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_produced_point"],
+        value_fn=lambda data: data["electricity_produced_point"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed_interval",
         translation_key="electricity_consumed_interval",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfEnergy.WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_interval"],
+        value_fn=lambda data: data["electricity_consumed_interval"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed_peak_interval",
         translation_key="electricity_consumed_peak_interval",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfEnergy.WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_peak_interval"],
+        value_fn=lambda data: data["electricity_consumed_peak_interval"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed_off_peak_interval",
         translation_key="electricity_consumed_off_peak_interval",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfEnergy.WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_off_peak_interval"],
+        value_fn=lambda data: data["electricity_consumed_off_peak_interval"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_produced_interval",
         translation_key="electricity_produced_interval",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfEnergy.WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_produced_interval"],
+        value_fn=lambda data: data["electricity_produced_interval"],
         entity_registry_enabled_default=False,
     ),
     PlugwiseSensorEntityDescription(
@@ -208,37 +215,35 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="electricity_produced_peak_interval",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfEnergy.WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_produced_peak_interval"],
+        value_fn=lambda data: data["electricity_produced_peak_interval"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_produced_off_peak_interval",
         translation_key="electricity_produced_off_peak_interval",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfEnergy.WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_produced_off_peak_interval"],
+        value_fn=lambda data: data["electricity_produced_off_peak_interval"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed_off_peak_point",
         translation_key="electricity_consumed_off_peak_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_off_peak_point"],
+        value_fn=lambda data: data["electricity_consumed_off_peak_point"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed_peak_point",
         translation_key="electricity_consumed_peak_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_peak_point"],
+        value_fn=lambda data: data["electricity_consumed_peak_point"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_consumed_off_peak_cumulative",
         translation_key="electricity_consumed_off_peak_cumulative",
         device_class=SensorDeviceClass.ENERGY,
         native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        value_fn=lambda data: data["sensors"][
-            "electricity_consumed_off_peak_cumulative"
-        ],
+        value_fn=lambda data: data["electricity_consumed_off_peak_cumulative"],
         state_class=SensorStateClass.TOTAL_INCREASING,
     ),
     PlugwiseSensorEntityDescription(
@@ -246,7 +251,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="electricity_consumed_peak_cumulative",
         device_class=SensorDeviceClass.ENERGY,
         native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_consumed_peak_cumulative"],
+        value_fn=lambda data: data["electricity_consumed_peak_cumulative"],
         state_class=SensorStateClass.TOTAL_INCREASING,
     ),
     PlugwiseSensorEntityDescription(
@@ -254,30 +259,28 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="electricity_produced_off_peak_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_produced_off_peak_point"],
+        value_fn=lambda data: data["electricity_produced_off_peak_point"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_produced_peak_point",
         translation_key="electricity_produced_peak_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_produced_peak_point"],
+        value_fn=lambda data: data["electricity_produced_peak_point"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_produced_off_peak_cumulative",
         translation_key="electricity_produced_off_peak_cumulative",
         device_class=SensorDeviceClass.ENERGY,
         native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        value_fn=lambda data: data["sensors"][
-            "electricity_produced_off_peak_cumulative"
-        ],
+        value_fn=lambda data: data["electricity_produced_off_peak_cumulative"],
         state_class=SensorStateClass.TOTAL_INCREASING,
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_produced_peak_cumulative",
         translation_key="electricity_produced_peak_cumulative",
         native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["electricity_produced_peak_cumulative"],
+        value_fn=lambda data: data["electricity_produced_peak_cumulative"],
         device_class=SensorDeviceClass.ENERGY,
         state_class=SensorStateClass.TOTAL_INCREASING,
     ),
@@ -287,49 +290,49 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         name="Electricity phase one consumed",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_phase_one_consumed"],
+        value_fn=lambda data: data["electricity_phase_one_consumed"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_phase_two_consumed",
         translation_key="electricity_phase_two_consumed",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_phase_two_consumed"],
+        value_fn=lambda data: data["electricity_phase_two_consumed"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_phase_three_consumed",
         translation_key="electricity_phase_three_consumed",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_phase_three_consumed"],
+        value_fn=lambda data: data["electricity_phase_three_consumed"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_phase_one_produced",
         translation_key="electricity_phase_one_produced",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_phase_one_produced"],
+        value_fn=lambda data: data["electricity_phase_one_produced"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_phase_two_produced",
         translation_key="electricity_phase_two_produced",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_phase_two_produced"],
+        value_fn=lambda data: data["electricity_phase_two_produced"],
     ),
     PlugwiseSensorEntityDescription(
         key="electricity_phase_three_produced",
         translation_key="electricity_phase_three_produced",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["electricity_phase_three_produced"],
+        value_fn=lambda data: data["electricity_phase_three_produced"],
     ),
     PlugwiseSensorEntityDescription(
         key="voltage_phase_one",
         translation_key="voltage_phase_one",
         device_class=SensorDeviceClass.VOLTAGE,
         native_unit_of_measurement=UnitOfElectricPotential.VOLT,
-        value_fn=lambda data: data["sensors"]["voltage_phase_one"],
+        value_fn=lambda data: data["voltage_phase_one"],
         entity_registry_enabled_default=False,
     ),
     PlugwiseSensorEntityDescription(
@@ -337,7 +340,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="voltage_phase_two",
         device_class=SensorDeviceClass.VOLTAGE,
         native_unit_of_measurement=UnitOfElectricPotential.VOLT,
-        value_fn=lambda data: data["sensors"]["voltage_phase_two"],
+        value_fn=lambda data: data["voltage_phase_two"],
         entity_registry_enabled_default=False,
     ),
     PlugwiseSensorEntityDescription(
@@ -345,7 +348,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="voltage_phase_three",
         device_class=SensorDeviceClass.VOLTAGE,
         native_unit_of_measurement=UnitOfElectricPotential.VOLT,
-        value_fn=lambda data: data["sensors"]["voltage_phase_three"],
+        value_fn=lambda data: data["voltage_phase_three"],
         entity_registry_enabled_default=False,
     ),
     PlugwiseSensorEntityDescription(
@@ -353,14 +356,14 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="gas_consumed_interval",
         icon="mdi:meter-gas",
         native_unit_of_measurement=UnitOfVolumeFlowRate.CUBIC_METERS_PER_HOUR,
-        value_fn=lambda data: data["sensors"]["gas_consumed_interval"],
+        value_fn=lambda data: data["gas_consumed_interval"],
     ),
     PlugwiseSensorEntityDescription(
         key="gas_consumed_cumulative",
         translation_key="gas_consumed_cumulative",
         device_class=SensorDeviceClass.GAS,
         native_unit_of_measurement=UnitOfVolume.CUBIC_METERS,
-        value_fn=lambda data: data["sensors"]["gas_consumed_cumulative"],
+        value_fn=lambda data: data["gas_consumed_cumulative"],
         state_class=SensorStateClass.TOTAL_INCREASING,
     ),
     PlugwiseSensorEntityDescription(
@@ -368,14 +371,14 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         translation_key="net_electricity_point",
         device_class=SensorDeviceClass.POWER,
         native_unit_of_measurement=UnitOfPower.WATT,
-        value_fn=lambda data: data["sensors"]["net_electricity_point"],
+        value_fn=lambda data: data["net_electricity_point"],
     ),
     PlugwiseSensorEntityDescription(
         key="net_electricity_cumulative",
         translation_key="net_electricity_cumulative",
         device_class=SensorDeviceClass.ENERGY,
         native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        value_fn=lambda data: data["sensors"]["net_electricity_cumulative"],
+        value_fn=lambda data: data["net_electricity_cumulative"],
         state_class=SensorStateClass.TOTAL,
     ),
     PlugwiseSensorEntityDescription(
@@ -384,21 +387,21 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.BATTERY,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=PERCENTAGE,
-        value_fn=lambda data: data["sensors"]["battery"],
+        value_fn=lambda data: data["battery"],
     ),
     PlugwiseSensorEntityDescription(
         key="illuminance",
         translation_key="illuminance",
         device_class=SensorDeviceClass.ILLUMINANCE,
         native_unit_of_measurement=LIGHT_LUX,
-        value_fn=lambda data: data["sensors"]["illuminance"],
+        value_fn=lambda data: data["illuminance"],
     ),
     PlugwiseSensorEntityDescription(
         key="modulation_level",
         translation_key="modulation_level",
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=PERCENTAGE,
-        value_fn=lambda data: data["sensors"]["modulation_level"],
+        value_fn=lambda data: data["modulation_level"],
         icon="mdi:percent",
     ),
     PlugwiseSensorEntityDescription(
@@ -407,7 +410,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         icon="mdi:valve",
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=PERCENTAGE,
-        value_fn=lambda data: data["sensors"]["valve_position"],
+        value_fn=lambda data: data["valve_position"],
     ),
     PlugwiseSensorEntityDescription(
         key="water_pressure",
@@ -415,14 +418,14 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.PRESSURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfPressure.BAR,
-        value_fn=lambda data: data["sensors"]["water_pressure"],
+        value_fn=lambda data: data["water_pressure"],
     ),
     PlugwiseSensorEntityDescription(
         key="relative_humidity",
         translation_key="relative_humidity",
         device_class=SensorDeviceClass.HUMIDITY,
         native_unit_of_measurement=PERCENTAGE,
-        value_fn=lambda data: data["sensors"]["humidity"],
+        value_fn=lambda data: data["humidity"],
     ),
     PlugwiseSensorEntityDescription(
         key="dhw_temperature",
@@ -430,7 +433,7 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["dhw_temperature"],
+        value_fn=lambda data: data["dhw_temperature"],
     ),
     PlugwiseSensorEntityDescription(
         key="domestic_hot_water_setpoint",
@@ -438,16 +441,16 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
         device_class=SensorDeviceClass.TEMPERATURE,
         entity_category=EntityCategory.DIAGNOSTIC,
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["sensors"]["domestic_hot_water_setpoint"],
+        value_fn=lambda data: data["domestic_hot_water_setpoint"],
     ),
-    PlugwiseSensorEntityDescription(
-        key="maximum_boiler_temperature",
-        translation_key="maximum_boiler_temperature",
-        device_class=SensorDeviceClass.TEMPERATURE,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        value_fn=lambda data: data["maximum_boiler_temperature"]["setpoint"],  # type: ignore [index]
-    ),
+    #    PlugwiseSensorEntityDescription(
+    #        key="maximum_boiler_temperature",
+    #        translation_key="maximum_boiler_temperature",
+    #        device_class=SensorDeviceClass.TEMPERATURE,
+    #        entity_category=EntityCategory.DIAGNOSTIC,
+    #        native_unit_of_measurement=UnitOfTemperature.CELSIUS,
+    #        value_fn=lambda data: data["maximum_boiler_temperature"]["setpoint"],  # type: ignore [index]
+    #    ),
 )
 
 PW_SWITCH_TYPES: tuple[PlugwiseSwitchEntityDescription, ...] = (
@@ -457,7 +460,7 @@ PW_SWITCH_TYPES: tuple[PlugwiseSwitchEntityDescription, ...] = (
         icon="mdi:water-plus",
         device_class=SwitchDeviceClass.SWITCH,
         entity_category=EntityCategory.CONFIG,
-        value_fn=lambda data: data["switches"]["dhw_cm_switch"],
+        value_fn=lambda data: data["dhw_cm_switch"],
     ),
     PlugwiseSwitchEntityDescription(
         key="lock",
@@ -465,13 +468,13 @@ PW_SWITCH_TYPES: tuple[PlugwiseSwitchEntityDescription, ...] = (
         icon="mdi:lock",
         device_class=SwitchDeviceClass.SWITCH,
         entity_category=EntityCategory.CONFIG,
-        value_fn=lambda data: data["switches"]["lock"],
+        value_fn=lambda data: data["lock"],
     ),
     PlugwiseSwitchEntityDescription(
         key="relay",
         translation_key="relay",
         device_class=SwitchDeviceClass.SWITCH,
-        value_fn=lambda data: data["switches"]["relay"],
+        value_fn=lambda data: data["relay"],
     ),
     PlugwiseSwitchEntityDescription(
         key="cooling_enabled",
@@ -479,7 +482,7 @@ PW_SWITCH_TYPES: tuple[PlugwiseSwitchEntityDescription, ...] = (
         icon="mdi:snowflake-thermometer",
         device_class=SwitchDeviceClass.SWITCH,
         entity_category=EntityCategory.CONFIG,
-        value_fn=lambda data: data["switches"]["cooling_ena_switch"],
+        value_fn=lambda data: data["cooling_ena_switch"],
     ),
 )
 
@@ -490,14 +493,14 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:hvac",
         icon_off="mdi:hvac-off",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["compressor_state"],
+        value_fn=lambda data: data["compressor_state"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="cooling_enabled",
         translation_key="cooling_enabled",
         icon="mdi:snowflake-thermometer",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["cooling_enabled"],
+        value_fn=lambda data: data["cooling_enabled"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="dhw_state",
@@ -505,7 +508,7 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:water-pump",
         icon_off="mdi:water-pump-off",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["dhw_state"],
+        value_fn=lambda data: data["dhw_state"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="flame_state",
@@ -513,7 +516,7 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:fire",
         icon_off="mdi:fire-off",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["flame_state"],
+        value_fn=lambda data: data["flame_state"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="heating_state",
@@ -521,7 +524,7 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:radiator",
         icon_off="mdi:radiator-off",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["heating_state"],
+        value_fn=lambda data: data["heating_state"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="cooling_state",
@@ -529,7 +532,7 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:snowflake",
         icon_off="mdi:snowflake-off",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["cooling_state"],
+        value_fn=lambda data: data["cooling_state"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="slave_boiler_state",
@@ -537,7 +540,7 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:fire",
         icon_off="mdi:circle-off-outline",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["slave_boiler_state"],
+        value_fn=lambda data: data["slave_boiler_state"],
     ),
     PlugwiseBinarySensorEntityDescription(
         key="plugwise_notification",
@@ -545,6 +548,6 @@ PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
         icon="mdi:mailbox-up-outline",
         icon_off="mdi:mailbox-outline",
         entity_category=EntityCategory.DIAGNOSTIC,
-        value_fn=lambda data: data["binary_sensors"]["plugwise_notification"],
+        value_fn=lambda data: data["plugwise_notification"],
     ),
 )