+++ custom_components/plugwise/sensor.py
@@ -28,10 +28,10 @@ async def async_setup_entry(
 
     entities: list[PlugwiseSensorEntity] = []
     for device_id, device in coordinator.data.devices.items():
-        if "sensors" not in device:
+        if not (sensors := device.get("sensors")):
             continue
         for description in PW_SENSOR_TYPES:
-            if description.key not in device["sensors"]:
+            if description.key not in sensors:
                 continue
             entities.append(
                 PlugwiseSensorEntity(
@@ -64,4 +64,4 @@ class PlugwiseSensorEntity(PlugwiseEntity, SensorEntity):
     @property
     def native_value(self) -> int | float:
         """Return the value reported by the sensor."""
-        return self.entity_description.value_fn(self.device)
+        return self.entity_description.value_fn(self.device["sensors"])